# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/src": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "src"
                ],
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dopTypes.PaginatedListRep"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entities.SrcSt"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dopTypes.ErrRep"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "src"
                ],
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/entities.SrcCUSt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dopTypes.CreateRep"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dopTypes.ErrRep"
                        }
                    }
                }
            }
        },
        "/src/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "src"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.SrcSt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dopTypes.ErrRep"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "src"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/entities.SrcCUSt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dopTypes.ErrRep"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "src"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dopTypes.ErrRep"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dopTypes.CreateRep": {
            "type": "object",
            "properties": {
                "id": {}
            }
        },
        "dopTypes.ErrRep": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "error_code": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "dopTypes.PaginatedListRep": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "results": {},
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "entities.SrcCUSt": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "notify_url": {
                    "type": "string"
                }
            }
        },
        "entities.SrcSt": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "notify_url": {
                    "type": "string"
                }
            }
        },
...
    }
}